Split `parseMessage` function to make it testable
Nuxt expects every file in the `api` directory to be a server function
so the test file needs to be in a different location. On top of that the
function itself can not be in the api code since it relies on Nuxt auto
imports which Vite does not support.